Remove clamping cursor to screen bounds
If we adjust the cursor coordinates to stay on the bounds of the screen,
then when the user moves their head to move the cursor off screen and
then back, the cursor will be shifted relative to their head position.

Instead, leave the cursor coordinates as they are (even if it would put
it off screen). Then, when the cursor is moved back onto screen, it
stays at the same orientation relative to the user's head.