Handle recalibration in main thread
Was experiencing intermittent failures to read input while attempting to
calibrate.

Originally, the RecalibrationController was placed in a new thread
because the main thread handled motion processing. Later, motion
processing was moved to a dedicated thread, while the main thread
continued running the app as long as the user did not quit. This was
implemented by reading from std::cin to check input, resulting in
intermittent missed input because of multithreaded std::cin accesses.

This commit restructures the app to handle recalibration in the main
thread, while the motion processing runs in the background.

Input handling and recalibration method call left in main(). To be
removed in a later commit.