Move input logic to a method in app class
After moving calibration input handling to the main thread, the logic
for this was left in main(). This commit moves this logic into a new
`runInteractive` method. This method starts the app, continues managing
user input while the user has not quit, starts the recalibration process
when needed, and stops the app after the loop ends (when getline fails
at EOF).

Having one public method with the logic to handle each stage of app
execution also allows for the `start` and `stop` methods to be made
private. These no longer need to be called outside the app class, as app
functionality is entirely handled in the new `runInteractive` method.

Since `stop` is now private, we can safely call it directly in the
destructor.

This separation can allow for implementing multiple run modes in the
future.

This resolves commit c8250d27