Move server to own thread; constantly update values
Instead of having the thread that runs the app logic telling the server
when to poll for the latest data (on that same app thread), the server
now runs on its own thread and continuously reads packets it receives
and stores the latest pitch and yaw values.

When the app thread wants to update the cursor, it calls the server's
`getLatest` method to retrieve the latest values.