Make server socket blocking
Now that server is on a dedicated thread, it's simpler and more
efficient to use blocking I/O while reading packet info from the socket
rather than use a sleep when we receive no data.

Because of this, we explicitly `shutdown` the socket when stopping the
server so that `recvfrom` unblocks and the thread exits and joins.

However, we do keep the small sleep in the App thread loop to pace the
cursor change events a bit.